feat: 친구 기록 화면 기능 구현 및 개선
- FriendRecord에서 LikeButton 표시 및 좋아요 토글 기능 구현
- CharacterInfoSection 더보기 메뉴 (차단하기) 구현
- WalkingResult에서 산책 완료 다이얼로그 표시
- Point API 및 Repository 구조 개선 (UserApi로 이동)
- DressRoom 다중 아이템 선택 및 장바구니 기능 구현
- DateUtils formatIsoToKorean 함수 날짜만 표시하도록 수정
- CartDialog 하단 표시 및 UI 개선
- FriendRecordViewModel에서 위치 정보 처리 개선

+++ app/src/main/java/team/swyp/sdu/data/remote/walking/dto/FollowerWalkRecordDto.kt
@@ -25,7 +25,7 @@ data class FollowerWalkRecordDto(
     val stepCount: Int = 0,
 
     @SerialName("totalDistance")
-    val totalDistance: Int = 0,
+    val totalDistance: Double = 0.0,
 
     @SerialName("likeCount")
     val likeCount: Int = 0,
